Subfield/Subindex Local Annotation Support  (#941)
This commit adds support for adding Subfield/Subindex local annotation , https://github.com/llvm/circt/issues/877.
Given any annotation on an array instance or subfield,  ``~Foo|Foo>bar[1].baz``,  ``circt::firrtl::fromJSON`` parses the target
string into a prefix(``~Foo|Foo>bar``) and suffix(``[1].baz``). The suffix being the subfield/subindex part. 
Here we assume the target specification are well formed and valid.
The suffix is then stored as part of the annotation, with the key "target", for example ``target = ["[1]", ".baz"]``, along with the actual annotation.
This suffix will be parsed by the lowerTypes, and the annotation will be applied on the appropriate lowered instance.
The lowerTypes, will construct the suffix from the String attribute array, and add the annotation to only those declarations after lowering which match this suffix.