[Handshake] Add basic System Verilog integration test. (#464)
This tests from the Standard dialect all the way to System Verilog
using Handshake. A simple function in the Standard dialect is given
and lowered through the Handshake flow.

The resulting design is tested by a System Verilog script that waits
for the design output to become ready, and prints it. This allows us
to FileCheck the output similarly to the handshake-runner tests.