[FIRRTL]  Add Port Direction (#992)
* [FIRRTL] Store module port direction as attribute

Store FIRRTL Dialect module/extmodule port directionality in a
"portDirections" integer attribute.  The bit position in the attribute
indicates directionality (a zero is an input, a one is an output).
Previously directionality was encoded as an outer flip on
ports.  (Output ports had an outer flip, input ports had no outer flip.)
Module ports are now verified to require that no outer flip is applied.

Directionality is printed/parsed as a leading "in" or "out" before the
port identifier.  E.g., a module with one input and one output now looks
like:

    firrtl.module(in %a: firrtl.uint<1>, out %b: firrtl.uint<1>)

Previously, this would look like:

    firrtl.module(%a: firrtl.uint<1>, %b: firrtl.flip<uint<1>>)

* [FIRRTL] Use Port Direction in LowerTypes
* [FIRRTL] Switch to Direction in ExpandWhens
* [FIRRTL] Use direction in BlackboxMemory Transform
* [FIRRTLToRTL] Use Port Direction
* [LLHD] Use FIRRTL port direction during conversion
* [Handshake] Use FIRRTL Directions in Conversion
* [FIRRTL] Fix tests for in/out direction, NFC

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>