[FIRRTL] Support bundle types in ExpandWhens
This adds bundle support to the firrtl-expand-whens pass.  This means
that it can track individual field's connections when declarations have
a bundle type.  Due to pass ordering, most declarations returning bundle
types have already been decomposed into multiple scalar declarations.
This is not the case with MemOp, which returns a bundle for each memory
port. Although the main target of this code change was MemOp
declarations, support for bundle types was easily extended to instances
and wires.  Registers with bundle types are not fully supported yet, as
they require an implicit connection to themselves.

In order to facilitate this feature, a new utility was added to allow
mapping to individual fields of a declaration. This new utility is named
SinkLocation, and it acts as a reference to a specific field in a bundle
or vector. Two different SinkLocations referring to the same field will
be equal and produce the same hash.

SinkLocations can be constructed from the result of a Subfield op, and
the constructor will crawl backwards through the IR to find the
declaration they are referring to.

SinkLocations can be extended to support fields in a vector, although
support for this was not needed for expand whens.

SinkLocations work by storing a vector of field indices which are a path
leading to the field they represent. Since bundle types with sink values
are relatively rare in the expand-whens pass, it makes sense to use a
map key which refers to specific fields. If a pass required storing
information about every field in a bundle, it might be beneficial to use
a denser data structure for the mapping.