[ExportVerilog] Pass an Operation to print type instead of Location.
Previously, the location was used to emit any errors. By passing an
Operation pointer, we still have the ability to do that, but it opens
up other possibillities. Specifically, we can use the Operation as an
anchor to lookup symbols, which is looking like it will be necessary
in order to emit types that reference type declarations.