Bind support
binds are remote instantiations of a module in another module.  This is modeled with bind.explicit, which takes the symbol of an instance.  The instance is marked with an attribute which prevents its verilog emission.  Where arguments to the instance are expressions, rather than things with names, temporary wires are created to give verilog-namable things to bind to.