[FIRRTL] Add new operation `BundleOp`
BundleOp is a new FIRRTL operation that takes several values and
combines them in to a single bundle.  This is primarily used to take the
multiple return values from an InstanceOp, and combine them in to a
single value which can be connected to or invalidated.

A follow on commit to this will simplify the handling of
emitInvalidate() in the FIRRTL parser.