[Python] Factor NamedValueBuilder out of InstanceBuilder, NFC. (#1132)
This adds a new base class to support named access to the operands and
results of an operation being built. Existing support in
InstanceBuilder is maintained and slightly cleaned up.

This sets the stage for other builders for ops that name their
operands and results, like seq.reg and a chunk of the comb dialect.