[FIRParser] Speed up translateLocation.  NFC
Lots of operations allow locators but generally don't have them, including
many expressions.  We currently plaster location nodes from the .fir file
itself, which is probably the wrong thing (should instead wait for the entire
statement to be parsed and take the locator from it).

Regardless, it doesn't make sense to continually unique the filename identifier
for the fir file itself, we can just cache it.