[HW] Add TypeAliasOr type, and use it to support ArrayType or alias.
The TypeAliasOr type represents either a type alias or an inner
type. This is to support operations that are defined in terms of a
concrete type, like ArrayType, rather than an abstract dialect type,
like HWValueType.

All array operations that previously expected ArrayType for operands
or results are updated to expect TypeAliasOr<ArrayType>. This required
some updates to parsers and ExportVerilog.