[FIRParser] Make the parser fold() the mux generated by invalidif.
This was noticed by inspection that it was missing after seeing a lot
of unfolded muxes getting generated.  Note that the consequence of this
is that there will be no unfolded muxes with invalid after the parser.

If we need to keep these around for some reason, then the MuxOp::fold
hook should be changed to not unconditionally fold these.