[FIRRTL] Support bundle types in ExpandWhens (#1045)
In order to facilitate this feature, a new utility was added to allow mapping to individual fields of a declaration. This new utility is named FieldRef, and it acts as a reference to a specific field in a bundle or vector. Two different FieldRefs referring to the same field will be equal and produce the same hash.

FieldRefs can be constructed from the result of a Subfield op, and the constructor will crawl backwards through the IR to find the declaration they are referring to.

FieldRefs can be extended to support fields in a vector, although support for this was not needed for expand whens.

FieldRefs work by storing a Value pointing to the operation which defined them, and a unique integer called the FieldID. The FieldID corresponds to the index of a field when making a pre-order depth-first walk of the bundle elements. The BundeType type was extended with some helpers to make FieldID calculation quick.