[FIRParser] Completely reimplement location processing in statements fixing Issue 1140
Previously we parsed subexpressions and gave them locations in the .fir file,
and did a best effort approach at constructing "subOps" which allowed us to
replace their locations with the the one specified in the locator.

This missed applying locations to some subexpressions, and also was inefficient
because it needed to create all the locators for the .fir file.  This wastes
time and memory.

This patch changes to a new approach, where we create the suboperations with
an invalid source locator, and listen to the OpBuilder to capture precisely
the set of operations created.  If a locator is found at the end of the
statement then it is applied to all the ops once without creating the
FileLineCol loc within the .fir file.  If there is no locator, we lazily
create those locations.

This patch completes that except for an issue with primitives which is still
eagerly calling translateLocation.  I'll deal with that as a follow on.