[SV] Use function for InterfaceOp builder body
Remove the default builders for sv::InterfaceOp and add a lone builder
that users a std::function<void()> to construct the body.  (This is
modeled off of how the always blocks work.)  This dramatically
simplifies the construction of interfaces and avoids boilerplate
whenever you want to place operations inside the interface.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>