[PyCDE] Add generator options to select a generator by name.
The `generate` call accepts a list of candidate generators to run,
which defaults to empty.

When looking up a generator, the list is consulted. If it is empty,
the first generator is chosen by default. If it is present, the first
generator in the list is selected. If multiple candidates are in the
list, an error is raised.