[FIRRTL][LowerTypes] Use FieldRefs instead of suffix strings (#1211)
This changes the way that lower types maps old bundle values to new
scalar values. The old way of mapping was using a pair of the bundle
value and a string representation of the field name:

```
(oldValue : Value, String fieldName : String) -> (newValue : Value)
```

The new mapping is using FieldRefs, which are a more efficient way to
represent the same thing. A FieldRef is a pair of an `oldValue`, and an
integer ID unique to the field.  The integer IDs are assigned to each
field in a bundle when walking it in a depth first (preorder) order.

This shaves off 9 seconds from LowerTypes on the generated "test3" test
case, bringing it from 60s to 51s.