[FIRRTL][LowerTypes] Convert partialconnects to connects
This modifies the lower tyes pass to canonicalize all partial connects
as regular connects.  This is something that happens in the SFC duing
the expand-connects pass, which is a separate earlier pass from
lower-types.

A partial connect has implicit truncation when the source value is
larger than the destination. When lower-types converts this to a regular
connect, it will emit a tail-op to truncate the source first.

Partial connects are partially common because the FIRRTL we are
consuming has implicit truncations in regular connect statements, which
we transform in to partial connects.

Removal of partial connects and in particular implicit truncation needs
to happen before the expand-whens pass runs.  The ExpandWhens pass will
try to combine multiple connects to a destination into a single connect
statement by muxing together the sources.  It would be tricky to have
expand whens properly truncate all source values to create the correct
mux operation.