[FIRRTL][ExpandWhens] Make sure values are passive before muxing
When a sink is used as a the RHS of a connect statement, the expand
whens pass may create a mux using that value.  The connect op can handle
operands with flip type, while the mux op cannot.  This checks if the
type is a flip and emits a the asPassive cast if it is not.