[FIRRTL][LowerTypes] Use a visitor pattern for flattening types
The flatten methods used to return a vector with an entry for each leaf
element in a type.  This changes the flatten method to be a "walk"-like
function that takes a function_ref and applied it to each leaf element.
By not building up the vectors we save a significant amount of
processing time.

This also allows the "suffix" string to be built lazily using twines
instead of creating many string, even when they were not used.