[FIRRTL] Cleanup asPassive casting in parser
Change all asPassive casting in the FIRRTL parser to introduce a cast if a value
is not passive or if it has sink flow.  This unifies some of the whack-a-mole
that was introduced in fa221e5a and continued in 3cdd619f.

Simplify the parser, by changing convertToPassive to also do flow checking as
defined above.  This then allows forcePassive (introduced in fa221e5a) to be
removed.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>