[PyCDE] Declare parameterized modules in functions (#1241)
Require parameters, inputs, and outputs to be declared before operation construction. The supported way is to declare the class in a function then return it. This ensures that all the parameters are in scope for the generators, so they no longer need to be passed into it. This is a rather large refactor but this new way of parameterizing modules makes more sense.