[FIRRTL][ExpandWhens] Split the module body processing
This change splits the way module bodies are processed from the way
WhenOp regions are processed.  There was unneeded work in the module
body to process simulation constructs, which is now skipped.

This creates two different visitor classes, one for the Module and one
for the WhenOp, and moves the common code to a new parent class.

This change also checks if any changes have been made to the module and
will `markAnalysisPreserved` if no changes have been made.