[FIRRTL][InferWidths] Allow connect truncation all the way to 0 bits (#1295)
The infer widths pass will emit a truncation when it needs to connect a
larger source to a smaller destination.  This code was not able to
handle the case when the destination was 0 bits.  This switched to using
the tail operation which is capable of truncating to a 0 bit value.