[FIRRTL] Fix issue #222: MLIR now supports locations on block arguments.
This adds location info to ModulePortInfo, preserves it in the parser, and
applies that loc info to the BlockArgument's of a module.  ExtModules do not
preserve this, but this could be added if anyone cares.  As far as I know,
Chisel doesn't provide useful location information for extmodule ports.

Previous commit wasn't correct.