[IMConstProp] Fix a race condition that magically popped up for me.
IMConstProp rewrites the modules in a circuit in parallel, and each
rewrite is intensely querying the latticeValue DenseMap.  As such,
we have to be careful to use read only accessors to the densemap,
but the subscript operator mutates when an entry is missing. Change
this to use the non-mutating find version of the algorithm.

I have no idea how to write a testcase for this sadly.