[FIRParser] Switch the per-module symbol table to a StringMap instead of ScopedHashTable. NFC.
.fir file semantics for scoping are interesting: when blocks are allowed
to reuse names in "then and else" blocks, but there is no shadowing allowed.
This makes ScopedHashTable overkill for our application.  Eliminating it
simplifies away the memory name handling logic, and allows us to use a
StringMap instead of using pre-uniqued mlir::Identifiers as the key to the
map.

The consequence of this is that we simplify the parser logic by quite a
bit - this eliminates the memoryScopeTable entirely (along with the
complicated logic to maintain it).  We also reduce pressure on the
MLIR identifier table, which probably deftly dodges Issue #1283.  This
also reduces contention on the Identifier table in multithreaded
mode, but this patch doesn't make any change related to that.

On my laptop, this reduces parser time on a large testcase from 17.4s
to 15.9s.