[ExportVerilog] Make alwaysff emit as always by default
Some of our tools disallow driving a value from both an `always_ff`
block and an `initial` block. The Verilog spec in section `9.2.2.4`
specifies that variables set in an `always_ff` cannot be be set by any
other process. If the initial block is considered a process, then our
verilog emission for this case is incorrect.

This commit changes the default lowering options to not emit
`always_ff`, which increases our tool compatability.  Down the road we
may change how we initialize variables, which could allow us to legally
emit `always_ff`.