[FIRRTL][ExpandWhens] Print one init error per run
When an array is not initialized, the expand whens pass will print an
error for every element in the array.  This could print far too many
errors on huge arrays which may be missing a single bulk connect.  Since
ExpandWhens is running after lower-types there doesn't seem to be a good
way to reduce the number of errors printed, for example printing 1 error
per sink in the original source.

This changes the pass to quit after finding the first initialization
error.