[HIR] Syntax changes in HIR.
Changed syntax of memref for banked dims. It is easier to understand
which dims are banked.

Now load/store op specifies port. Port is kept optional so that an
optimization pass can fill in the port later.

Now LoadOp specifies delay.

Moved ports for memref and bus types outside the type. The LoadOp does
not need to know how many ports a memref has!