[ExportVerilog] Extern module's port types cannot be trusted, use value types
External modules' types are not necessarily the same for every instance due to
SV module parameterization. Thus, we cannot rely on the extern module's port
declaration to determine to print a port or not. We can rely on the instance
operations' operand types in both cases since a verifier ensures that the types
match in the internal module case.