Attempt to avoid re-naming when unnecessary.
This will skip renaming for: 1) the first time a module with
parameters is instantiated, and 2) modules that are unparameterized.

To be able to fully skip renaming if all instantiations of a
parameterized module are created, we'll either need some global data
on the side, or start putting the parameters into the HWModuleOp.