Add some more FIRRTL annotation documentation (#1330)
This moves the annotation documentation out of the FIRRTL rational and in to its own document. It expands the target section a little bit with some details from a Google Docs document. At the end of the document is reference style material for each annotation. This works pretty well for the simple annotations, but for GrandCentral each annotation had to have several tables to describe the format. Details were collected from the Scala source code and our tests.

The Scala format specifier was previously used to get decent highlighting for FIRRTL code, but this does not render well on the CIRCT website.