This adds the ability to have constant values of type `Clock`, `Reset`, and `AsyncReset`.
The SSA value format for constant op has been extended to handle the new
return types:
```
%c1_clock = firrtl.constant 1 : !firrtl.clock
%c1_reset = firrtl.constant 1 : !firrtl.reset
%c1_asyncreset = firrtl.constant 1 : !firrtl.asyncreset
```

New folders for `asClock()` and `asAsyncReset()` have been added to
transform `cast(constant) -> constant`.

This fixes a bug in IMConstProp, where it was incorrectly attempting to
materialize a constant of type `firrtl.reset` due to an implicit
conversion of a `UInt<1>` from a module instantiation to the module
body.