[ExportVerilog] Add naive BindInterfaceOp emission
Add basic emission of BindOpInterface.  This emits using the janky
"(.*)" approach (where anything matching is connected).  This is
non-ideal, but interfaces are not as straightforward to handle as
instances where all connections are explicit.

This is intended to be a temporary emission strategy that can be
improved in the future.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>