[Calyx] Add Go Insertion Pass, GroupGoOp. (#1387)
Adds the "Go Insertion" pass and GroupGoOp. This is first of 3 passes (GoInsertion, 
CompileControl, RemoveGroups) necessary to "flatten" the Calyx IR before lowering 
to RTL (more details in the paper). This pass inserts guards associated with the group
to each assignment within the group.

Adds a GroupGoOp, and creates a base class for ports associated with GroupOps.