[CMake] Fix incorrect file dependencies for configured files
In a couple of places in our build system, we wanted to create targets
that depended on generated files. We were not passing the full path to
the generated files, e.g. `CosimScheme.h`, and CMake was unable to find
the file. CMake was searching then searching the current directory for a
matching file with a file prefix, and CMake would find the unprocessed
file, e.g. `CosimSchema.h.in`, and use that instead.  A policy change
CMP0115 in newer version of CMake causes this to spit out many warnings
about it how it will no longer search for source files. For more
information about the policy see
https://cmake.org/cmake/help/latest/policy/CMP0115.html

This change provides the full path to the generated file, which allows
cmake to find the intended file and not produce so many warnings.