[HIR] Added ForOp latched inputs.
ForOp can body can not directly capture any value that is associated with a
time instant from a parent region. These values need to be passed to the
body via latch inputs.
At the beginning of a ForOp all these values are latched for the
duration of the loop and can be used at any time instant in the loop
body (like constants).
This helps in simplifying the RegisterAllocationPass - otherwise finding
a relation between the outer region's time-var and the loop's
iter-time-var becomes tricky.