[FIRRTL] Fix GCT Data Taps for ExtModule Sources
Fix GrandCentral Annotation scattering and GrandCentralTaps pass to
support external module _sources_.  Previously, it was assumed that an
external module port had to be a "portName" sink.  However, it is
entirely possible for an external module port to be tapped.

Implement this fix with the following changes:

First, a new "type" field that can be either "portName" or "source" is
added during scattering of ReferenceDataTapKey Annotations.  This
unambiguously marks these taps as either a sink or a source.

Second, the map used to store tap points for module ports (by tapping
BlockArguments) is changed to work generically on _ports_ of modules or
extmodules.  (This is necessary because extmodules do not have block
arguments).

Third, the key used for each map is changed to use an <id, portID> tuple
for a key.  Previously, this would rely on the actual Annotation as the
key (which was effectively encoding a superset of the same information,
id and portID, as well as useless information about the annotation
name).  Changing annotations to encode type information ("type" of
"portName" or "source") means that sources and sinks do _not_ have the
exact same annotation as before and only the common information (id and
portID) should be used to connect them.

Update the tests to include "type" information.  Add a test of tapping
an external module port.

Co-authored-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>
Co-authored-by: Fabian Schuiki <fabian@schuiki.ch>
Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>