[FIRRTL] Try to get windows compiling (#1430)
On our Windows builder configuration, MSVC is failing to token paste `add`,
`not`, `or`, and `xor` when expanding the `TOK_LPKEYWORD` macro. The macro
`TOK_LPKEYWORD_PRIM` normally delegates to `TOK_LPKEYWORD` when it is not
defined by the user.  Manually definining `TOK_LPKEYWORD_PRIM` to the same
definition as `LP_KEYWORD` allows the build to pass.

I found this by deleting code until I found the problematic block, and then
tried a couple different things to get it to compile. I tried looking through
the patches from June 21-22 2021 (when the Windows builds broke) to spot an
issue, but nothing jumped out at me.

This block of code has been here since aab88c0 from Apr 5 2020, the most recent
relevant change appears to be 17f25c7 from May 25 2021. There is a similar (but
not identical) use of `TOK_LPKEYWORD` in the lexer with no problems.