[Calyx] Add helper functions, other cleanup for upcoming CompileControl pass. (#1404)
Cleans up the Calyx dialect a bit in preparation for the Compile Control pass. Removes HW 
dialect from GoInsertion pass; it is unused. Add a component builder that uses 
ComponentPortInfo to define ports. Add helper functions to access the GoOp 
and DoneOp of a group. In the GoInsertion pass, place the calyx.undef value 
in the calyx.wires section. Remove the result from the DoneOp.