[Python] Drop reference to Edge instance after erasing it.
We keep a dict of backedges on OpViews that need them, so they may be
cleaned up later. In order to print the OpView in a situation where
the Edge was _not_ cleaned up, we store a reference to the OpView in
the Edge. This is fine, but in the event that the Edge _is_ cleaned
up, we were leaving the Edge in this dict of backedges, which meant
there was still a Python-side reference to the OpView.

Down the road, that prevents the OpView's destructor from running and
removing it from the liveOperations map at the appropriate time. By
forcibly dropping the Edge from the dict of backedges, we allow the
liveOperations map to be kept in sync.

The small change to the existing test case is enough to trigger the
problematic behavior and demonstrate the issue.