[FIRRTL] Add async reset const check to LowerToHW
When lowering a `firrtl.regreset` with async reset in `LowerToHW`, check
that the reset value is a constant. This is a common constraint imposed
by EDA software and is enforced on the Scala side as well (in the
CheckResets pass).

This also finally enables the `async-reset-errors.mlir` integration test
that we had for quite a while now.