Use mlir::parallel*, not llvm::parallel*, NFC
Switch existing usages of llvm::parallel* methods to provide parallelism
within a pass to use the newer mlir::parallel* methods.  The MLIR
variants avoid the need for explicitly checking if parallelism is
allowed by the context.

There is no upstream equivalent of llvm::parallelTransformReduce, so the
one usage of that in LowerToHW remains in place.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>