[FIRRTL] Fix Annotation LowerTypes SubAnno Bug
Fix a bug where an subAnno straddles a range of field IDs which cross
what would traditionally be a field boundary.  I.e., subAnnos are
expressive enough to cover any contiguous range of field IDs.  This did
not work correctly in lowering and this changes it to do what you would
think: subAnnos are split and applied to whatever field IDs they
originally applied to.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>