[FIRRTL] Error on Annotations that Do Not Apply (#1440)
Change the FIRRTL parser to error if any Annotations are unapplied to
the circuit.  Put differently, this will error if any Target in an
Annotation (or a generated Target from Annotation scattering) points at
something which does not exist in the circuit.

Residual Annotations indicate either a user problem like inconsistent
FIRRTL and Annotation files or some problem with scattering logic or the
parser itself.  E.g., you can hit this if you have busted scattering
code (that creates targets that do not exist) or for cases where
non-local Annotations get through scattering logic.  This is a situation
that needs to be addressed and continuing compilation may result in
extremely tedious errors due to the compiler effectively dropping
Annotation information which is assumed to be critical to control
compilation in a way the user wants.

Together with a894ad8c and a08f15ca this should cause all unapplied
Annotations to produce parse-time errors.

Add tests to check if annotations that do not apply to anything in a
circuit expectedly throw errors.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>