[FIRRTL][Annotation] Remove maxFieldID param from SubAnnoAttr (#1462)
Previously, `SubAnnoAttr` had a `minFieldID` and a `maxFieldID` to represent the
range of `fieldID`s it is targeting. The `maxFieldID` is introduced for the
`LowerTypes_v1` to easily tell whether a leaf element is targeted by a
`SubAnnoAttr` when filtering annotations without recursively looking into the
aggregate type.

However, this `fieldID` range introduced many undefined things like cross-fields
annotations (which are not allowed), making `SubAnnoAttr` hard to verify and
handle. Meanwhile, as `LowerTypes_v2` doesn't flatten aggregate types at once,
but recursively peel off the outermost layer of aggregate types, it no longer
needs to answer the question above. Therefore, we decide to make `maxFieldID`
retire.

This patch removed `maxFieldID` from `SubAnnoAttr` and renamed `minFieldID` to
`fieldID`. Related codes in `FIRParser` and `LowerTypes` are updated as well.

- Fixes #1453.
- Closes #1457.