[FIRRTL][GCT] Generate Binds for Grand Central Interfaces (#1434)
* [FIRRTL][GCT] Scatter name, defname on parent

Scatter the "defName" and "name" fields from a ViewAnnotation into the
existing Grand Central (GCT) attribute on a FIRRTL module.  This
information is necessary to know which interface (identified by
"defName") should be bound into the module with what instantiation
name (identified by "name").

* [LowerToHW] Allow sv::InterfaceOps to pass through

Add support for operations of specific types to pass through FIRRTL to
HW lowering.  This uses a vector of operations which are inserted into
the outer module after FIRRTL modules are lowered.

Use this to allow SystemVerilog InterfaceOps to pass through.
Previously, these would cause an error and get dropped.

This is done to allow for FIRRTL transforms to add operations which may
have no representation in the FIRRTL dialect.  This is viewed as better
than, alternatively, encoding operations in annotations/attributes and
having LowerToHW expand them.

* [FIRRTL][GCT] Instantiate and bind interfaces

Change GrandCentral to add an interface instance of the
annotation-encoded instance inside the "parent" FIRRTL module.  Place a
bind in the outer module (which is outside the circuit).

* [FIRRTL][GCT] Test interface binding behavior, NFC

Add a test that interface binding works as expected for code coming out
of the GrandCentral pass.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>