[LowerToHW] Add whitelist for remaining annotation warnings
Certain annotations are expected to remain in the IR until LowerToHW
runs. For example, annotations that prevent a pass from running on an
operation have no effect if the pass doesn't run at all. But in that
case, the pass also hasn't consume the annotations. This includes things
like don't-obfuscate and don't-dedup.

This PR adds a whitelist to `warnOnRemainingAnnotations` in the
`LowerToHW` pass to allow a select set of annotations to be silently
dropped.