[FIRRTL] Invalidate Module Outputs in IMCP
Invalidate all ground-type module outputs (which are not the top module)
for circuits where the top module is a module.  This is done to treat
these ports (which are sinks/lvalues) like wires when they are used as
RHS sources/rvalues.  All aggregate module outputs are marked
overdefined.

This blocks removal of the register "r" in the following FIRRTL
module (when this module is NOT the top module).  Without this patch,
output "a" never receives a lattice value:

    module Foo:
      input clock: Clock
      output a: UInt<1>

      reg r: UInt<1>, clock
      r <= not(r)
      a <= r

This is a follow-up to 171dc4cf which got this working for this register
formulation in the top module.

Fixes #1478.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>