[PyCDE] Re-name types->pycde_types, NFC.
I once again ran into issues with circular imports related to the
pycde.types module. This time I dug into it a bit, and I think the
issue is that there is a builtin types module in Python. If, in the
process of importing pycde.types, anything imports types, there will
be an error related to importing a partially initialized module.

This just renames pycde.types to the redundant pycde.pycde_types to be
sure there is no conflict with the builtin types module.