[FIRRTL] Update output ports in IMCP lattice (#1490)
Change IMCP to update output ports and not just _external_ users of
output ports.  Output ports have wire-like semantics and may be read
from as well as written to.

Add a test that IMCP doesn't remove registers which are illegal to
remove.  This is testing both RegOp and RegResetOp across a wide array
of different formulations which tickle weird behavior in IMCP.

Add a test that IMCP works for RHS output port.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>