Revert "[FIRRTL][LowerToHW] Handle 0-bit mux sel (#1349)"
Revert a984a22ff22bd11460faea34dafb1efbe1e43e4a.

This reverts changes that allowed a FIRRTL mux with a 0-bit condition to
both be constructed and folded to the false condition.  The consensus is
that this is something that should _not_ be allowed in the FIRRTL spec
and should be an error.

Fixes #1471.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>